Preserve presenter regardless of view lifecycle
