--- app/src/main/java/com/patrickgatewood/weather/ui/PresenterOps.java
@@ -1,8 +1,0 @@
-package com.patrickgatewood.weather.ui;
-
-
-public interface PresenterOps {
-
-    void onFetchButtonClick();
-    void onResume();
-}
