+++ app/src/main/java/com/patrickgatewood/weather/ui/WeatherView.java
@@ -1,6 +1,6 @@
 package com.patrickgatewood.weather.ui;
 
 
-interface ViewOps {
+interface WeatherView {
     void updateTextViews(String temperature, String feelsLikeTemp, String summary);
 }
